refactor(pen): systems/ を canvas.worker.ts へ集約 + Ubi.state.topLevel 採用
- pen:canvas: PenWatchSystem / PenInputSystem / PenSyncSystem / PenCanvasSystem
  4 ファイルを canvas.worker.ts に統合。penFingerprint.ts も同ファイルへ。
  "1 Component = 1 ファイル" を徹底し、見通しが大きく改善。
- pen:pen: lockedBy を手動 reducer で追跡していたが、Ubi.state.topLevel<string|null>(null)
  に置き換え。PenSelfSystem と local オブジェクトを撤去。`selectMe` / `releaseMe` は
  `pen.local.lockedBy = ...` への代入だけで自動同期。
