--- plugins/pen/src/systems/index.ts
@@ -1,8 +1,0 @@
-/**
- * Pen Plugin - System Exports
- */
-
-export { PenCanvasSystem } from './PenCanvasSystem';
-export { PenInputSystem } from './PenInputSystem';
-export { PenSyncSystem } from './PenSyncSystem';
-export { PenWatchSystem } from './PenWatchSystem';
